+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.2
 
 - bump mock-fs dependency version to fix babel-polyfill problems running the test
+- Add support for the architectures multi-value property
 ## 0.1.1
 
 - Allow 'special' characters in the library name (such as +)